ams-qd:  Use modulo, not remainder, in Random()

+++ v68k/modules/ams-qd/QDUtils.cc
@@ -18,7 +18,7 @@ pascal short Random_patch()
 {
 	QDGlobals& qd = get_QDGlobals();
 	
-	qd.randSeed = qd.randSeed * 16807 % 0x7FFFFFFF;
+	qd.randSeed = UInt32( qd.randSeed * 16807 ) % 0x7FFFFFFF;
 	
 	return qd.randSeed;
 }
